requirements.txt
Django==5.0.3
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.28.6
dj_database_url==2.1.0
django-environ==0.11.2
django-filter==24.1
psycopg2==2.9.9